+++ HenriksHobbyLager/Repositories/Crud/DeleteProduct.cs
@@ -2,18 +2,18 @@
 
 public class DeleteProduct
 {
-    Console.WriteLine("Skriv in produkt-id: ");
+    // Console.WriteLine("Skriv in pr odukt-id: ");
     // TODO: Look this over.
     // if (!int.TryParse(Console.ReadLine(), out int id))
     // {
     //     Console.WriteLine("Ogiltigt ID! Bara siffror är tillåtna här.");
     //     return;
     // }
-    var productId = int.Parse(Console.ReadLine());
+    // var productId = int.Parse(Console.ReadLine());
     // if (product == null)
     // {
     //     Console.WriteLine("Produkt hittades inte! Puh, inget blev raderat av misstag!");
     //     return;
     // }
-    _productFacade.DeleteProduct(productId);
+    // _productFacade.DeleteProduct(productId);
 }
